requirements.txt
asgiref==3.8.1
defusedxml==0.7.1
dj-rest-auth==6.0.0
Django==5.0.7
django-hyperview==0.0.2
djangorestframework==3.15.2
djangorestframework-xml==2.0.0
pillow==10.4.0
psycopg2-binary==2.9.9
sqlparse==0.5.0
